# Conflicts:
#	features/discovery/home/src/main/java/com/tokopedia/home/beranda/domain/model/DynamicHomeChannel.kt
#	features/discovery/home/src/main/java/com/tokopedia/home/beranda/presentation/view/adapter/factory/HomeAdapterFactory.kt
#	features/discovery/home_component/src/main/java/com/tokopedia/home_component/HomeComponentTypeFactory.kt
#	features/recharge/digital_homepage/src/main/java/com/tokopedia/digital/home/presentation/adapter/RechargeHomepageAdapterTypeFactory.kt
#	features/tokopedianow/src/main/java/com/tokopedia/tokopedianow/home/presentation/adapter/HomeAdapterTypeFactory.kt
#	testapp/src/main/java/com/tokopedia/tkpd/MainActivity.kt

+++ features/logistic/logisticCommon/src/main/java/com/tokopedia/logisticCommon/data/model/CustomProductLogisticModel.kt
@@ -8,7 +8,7 @@ data class CustomProductLogisticModel (
 data class CPLProductModel(
     var productId: Long = 0,
     var cplStatus: Int = 0,
-    var shipperServices: List<Int> = listOf()
+    var shipperServices: List<Long> = listOf()
 )
 
 data class ShipperListCPLModel(
@@ -18,7 +18,7 @@ data class ShipperListCPLModel(
 )
 
 data class ShipperCPLModel(
-    var shipperId: Int = 0,
+    var shipperId: Long = 0,
     var shipperName: String = "",
     var logo: String = "",
     var shipperProduct: List<ShipperProductCPLModel> = listOf(),
@@ -26,7 +26,7 @@ data class ShipperCPLModel(
 )
 
 data class ShipperProductCPLModel(
-    var shipperProductId: Int = 0,
+    var shipperProductId: Long = 0,
     var shipperProductName: String = "",
     var uiHidden: Boolean = false,
     var isActive: Boolean = false
